Fixed and made it runnable
