--- src/main/java/com/aruna/example/SpringMVC/ServletInitializer.java
@@ -1,13 +1,0 @@
-package com.aruna.example.SpringMVC;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(SpringMvcExampleApplication.class);
-	}
-
-}
