+++ src/main/java/com/example/init/Application.java
@@ -1,0 +1,19 @@
+package com.example.init;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.*;
+import org.springframework.boot.web.support.*;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@SpringBootApplication
+@ComponentScan({"com.example.controller", "com.example.domain", "com.example.init"})
+@Configuration
+public class Application extends SpringBootServletInitializer{
+
+	public static void main(String[] args) {
+        new Application()
+                .configure(new SpringApplicationBuilder(Application.class))
+                .run(args);
+	}
+}
