--- src/test/java/com/aruna/example/SpringMVC/SpringMvcExampleApplicationTests.java
@@ -1,16 +1,0 @@
-package com.aruna.example.SpringMVC;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SpringMvcExampleApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
