+++ src/test/java/com/example/springmvc/ApplicationTests.java
@@ -1,0 +1,87 @@
+package com.example.springmvc;
+
+import java.io.*;
+import java.net.*;
+import org.junit.*;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class) @SpringBootTest public class ApplicationTests {
+
+    @Test public void getProducts() {
+
+        try {
+
+            URL url = new URL("http://localhost:8080/json/product/get");
+            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
+            conn.setRequestMethod("GET");
+            conn.setRequestProperty("Accept", "application/json");
+
+            if (conn.getResponseCode() != 200) {
+                throw new RuntimeException("Failed : HTTP error code : " + conn.getResponseCode());
+            }
+
+            BufferedReader br = new BufferedReader(new InputStreamReader((conn.getInputStream())));
+
+            String output;
+            System.out.println("Output from Server .... \n");
+            while ((output = br.readLine()) != null) {
+
+                System.out.println(output);
+            }
+
+            conn.disconnect();
+
+        } catch (MalformedURLException e) {
+            e.printStackTrace();
+            Assert.assertTrue("Should not get an exception", false);
+        } catch (IOException e) {
+            e.printStackTrace();
+            Assert.assertTrue("Should not get an exception", false);
+
+        }
+    }
+
+    @Test public void postProducts() {
+        try {
+
+            URL url = new URL("http://localhost:8080/json/product/post");
+            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
+            conn.setDoOutput(true);
+            conn.setRequestMethod("POST");
+            conn.setRequestProperty("Content-Type", "application/json");
+
+            String input = "{\"qty\":100,\"name\":\"iPad 4\"}";
+
+            OutputStream os = conn.getOutputStream();
+            os.write(input.getBytes());
+            os.flush();
+
+            if (conn.getResponseCode() != HttpURLConnection.HTTP_CREATED) {
+                throw new RuntimeException("Failed : HTTP error code : " + conn.getResponseCode());
+            }
+
+            BufferedReader br = new BufferedReader(new InputStreamReader((conn.getInputStream())));
+
+            String output;
+            System.out.println("Output from Server .... \n");
+            while ((output = br.readLine()) != null) {
+
+                System.out.println(output);
+            }
+            conn.disconnect();
+
+        } catch (MalformedURLException e) {
+            e.printStackTrace();
+            Assert.assertTrue("Should not get an exception", false);
+        } catch (IOException e) {
+            e.printStackTrace();
+            Assert.assertTrue("Should not get an exception", false);
+        }
+
+    }
+
+}
+
+
